add create batch

+++ app/api/v1/endpoints/schedule.py
@@ -80,31 +80,70 @@ async def post_schedule_batch(
     category_service: Annotated[CategoryService, Depends()],
     current_user_id: int = Depends(get_current_user_id)
 ):
+    """
+    tableName	“events” | “tasks” | “routines” | “categories”	어떠한 모델에 대한 수정인지
+    tempId	string	기존의 임시 ID
+    serverId	string	새로 생성된 서버의 ID
+    """
+    mapping_list = [] # response로 반환할 매핑 리스트
+
     for operation in request.operations:
         if operation.table_name == "events":
-            if operation.operation == "update":
+            if operation.operation == "create":
+                event_create_request = event_schemas.EventCreateRequest.model_validate(operation.payload)
+                real_event = await event_service.create_event(event_create_request, current_user_id)
+                mapping_list.append({"tableName": "events", "tempId": operation.row_id, "serverId": real_event.id})
+            elif operation.operation == "update":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "events"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "events")
                 event_edit_request = event_schemas.EventEditRequest.model_validate(operation.payload)
                 await event_service.edit_event(operation.row_id, event_edit_request, current_user_id)
             elif operation.operation == "delete":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "events"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "events")
                 await event_service.delete_event(operation.row_id, current_user_id)
         elif operation.table_name == "tasks":
-            if operation.operation == "update":
+            if operation.operation == "create":
+                task_create_request = task_schemas.TaskCreateRequest.model_validate(operation.payload)
+                real_task = await task_service.create_task(task_create_request, current_user_id)
+                mapping_list.append({"tableName": "tasks", "tempId": operation.row_id, "serverId": real_task.id})
+            elif operation.operation == "update":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "tasks"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "tasks")
                 task_edit_request = task_schemas.TaskEditRequest.model_validate(operation.payload)
                 await task_service.edit_task(operation.row_id, task_edit_request, current_user_id)
             elif operation.operation == "delete":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "tasks"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "tasks")
                 await task_service.delete_task(operation.row_id, current_user_id)
         elif operation.table_name == "routines":
-            if operation.operation == "update":
+            if operation.operation == "create":
+                routine_create_request = routine_schemas.RoutineCreateRequest.model_validate(operation.payload)
+                real_routine = await routine_service.create_routine(routine_create_request, current_user_id)
+                mapping_list.append({"tableName": "routines", "tempId": operation.row_id, "serverId": real_routine.id})
+            elif operation.operation == "update":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "routines"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "routines")
                 routine_edit_request = routine_schemas.RoutineCreateRequest.model_validate(operation.payload)
                 await routine_service.update_routine(operation.row_id, routine_edit_request, current_user_id)
             elif operation.operation == "delete":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "routines"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "routines")
                 await routine_service.delete_routine(operation.row_id, current_user_id)
         elif operation.table_name == "categories":
-            if operation.operation == "update":
+            if operation.operation == "create":
+                category_create_request = category_schemas.CategoryCreate.model_validate(operation.payload)
+                real_category = await category_service.create_category(category_create_request, current_user_id)
+                mapping_list.append({"tableName": "categories", "tempId": operation.row_id, "serverId": real_category.id})
+            elif operation.operation == "update":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "categories"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "categories")
                 category_edit_request = category_schemas.CategoryUpdate.model_validate(operation.payload)
                 await category_service.update_category(operation.row_id, category_edit_request, current_user_id)
             elif operation.operation == "delete":
+                if operation.row_id in [item["tempId"] for item in mapping_list if item["tableName"] == "categories"]:
+                    operation.row_id = next(item["serverId"] for item in mapping_list if item["tempId"] == operation.row_id and item["tableName"] == "categories")
                 await category_service.delete_category(operation.row_id, current_user_id)
 
 
-    return ApiResponse(message="일정 변경사항을 일괄 배치하였습니다.", data=None)
+    return ApiResponse(message="일정 변경사항을 일괄 배치하였습니다.", data={"results": mapping_list})
